Restrict Cloudflare protected HAProxy services to fixed IP set
The Cloudflare WAF setup requires that you restrict your service to only
their proxy IP ranges. This change extends our HAProxy service
definition to accommodate this by using a new ACL and an http-request
rule when the service has the "cloudflare_protected" flag.

This is not the only design option, but we have not yet needed to extend
our services with arbitrary rules, so it did not seem to be the time to
do that.

The IP list is inlined here, but the file could be placed by template,
from fileserver, or anything else (e.g., a custom deb package). The IPv4
ranges have not changed since September 2021.

See also: https://www.cloudflare.com/ips/