Switch mod_remoteip parameters to use upstream names
The puppetlabs/apache module has deprecated the proxy_ips parameter we
have been using, in favor of more specific parameter names that align
with the actual Apache/mod_remoteip directives. These are
RemoteIPTrustedProxy and RemoteIPInternalProxy, respectively.

Each of these behaves the same, except that "trusted" proxies are
designed to refer to external proxies. Client IPs reported from these
within ARIN-designated private networks are rejected (not promoted from
header to request attribute), since they would be invalid (non-internet
addresses reported on internet requests). An "internal" proxy is allowed
to report any address as the client IP.

Internally, the module was mapping to "internal" when proxy_ips was
specified. Our HAProxy servers are appropriate for this designation, so
this commit switches to the explicit parameter name, internal_proxy.