Use Cloudflare headers to set X-Client-IP
When a service is directly exposed, we set the X-Client-IP to the
connecting IP address. When proxying through Cloudflare, the originating
IP address is set in the CF-Connecting-IP header, which we should use
for the value instead.

This change also conducts a minor reordering of the template to keep the
forwarded protocol next to the client IP header directive. This
reordering is cosmetic only; the routing/access decisions are unchanged.