Support rolling deployment of Cloudflare WAF
By separating the connecting IP handling from the access restriction, we
allow a period where direct and proxied clients can access a service.
After we are satisfied that all DNS propagation is complete, we can then
restrict to proxied clients only.

The strategy here involves unconditionally checking the source IP and
whether the CF-Connecting-IP header is supplied. We unconditionally set
the X-Client-IP to the source. Then, we conditionally overwrite it with
the supplied header value, if and only if its is supplied and the source
is a Cloudflare proxy server. The access restriction is only enforced
when the service's cloudflare_protected property is true.