Restart every VM after a crash
The bit of code that configures VMs to autostart was inside the "make
sure this VM isn't already installed" condition, which was either a
mistake or a shortest path to get the tests to pass. I've fixed the
tests, so this commit will also mark every VM it manages as autostart.

Otherwise, this will only set a VM to restart after a crash if it is
currently set to destroy after a crash.

All the options for what to do after a crash are:

1.  destroy
2.  restart
3.  rename-restart
4.  preserve
5.  coredump-destroy (only available on crash, not e.g. poweroff)
6.  coredump-restart (only available on crash, not e.g. poweroff)

I haven't looked into what any of these mean, but coredump is an
intriguing word. This commit simply changes anything that's option 1 to
option 2.