HathiTrust TTO-288: send shib headers to Dex even when blank
Currently, HathiTrust uses a forked version of Dex
(https://github.com/hathitrust/dex-shib-proxy) to present itself as an
OIDC identity provider, serving as a proxy to the variety of SAML
identity providers that HathiTrust supports. This forked version
supports using additional headers to populate various claims in Dex.

As of Dex 2.38, Dex supports additional claims via the auth proxy.
However, if one of the headers it is using is missing, Dex will report
an error. This changes the behavior to always send the headers, even
when empty.

Additionally, this version of Dex expects the delimiter for the 'groups'
claim to be "," rather than ";".