Debug remote_write by limiting to recent samples
We're running into a lot of errors stemming from old data retries while
establishing the baseline thresholds and resource requirements for
Mimir, so this limits the samples to send via remote_write to only those
from the past 5 minutes.

We will want to increase this later, possibly to the default of
including everything in each TSDB (configured as 0 seconds).