Add etcd db size alerts
EtcdDatabaseFillingUpFast should probably be `severity: page`, but I
don't want to do that without letting this live for a while.

BigEtcdDatabase exists with the intention that alerts shouldn't stop
when the db finishes running out of space, but that amount of space is
2Gi, and I'm setting this alert at 500Mi because I've never seen a
healthy etcd db go above 500. Does that ever happen? I'd like to know.

We can tune it as needed when we know more.