Manage ownership and permissions for prometheus tls files
The initial motivation was simply to set the mode to 600 instead of
644 for the key files, but prometheus runs as `nobody:nogroup` so this
necessitated also managing ownership. To be consistent, we're also
managing that for the nonsecret files and setting 644 explicitly to
keep it clear.

Further, in testing, we noticed that updating certs failed to restart
prometheus, so we added `notify` to these files so prometheus will
actually use the new files whenever they change (as is already happening
with other config files).