start puppet if it's stopped
Currently it's easy to run `puppet agent -t` and assume that the puppet
agent service is running when it's not (esp. during puppet installation
and upgrade).

When we want to disable puppet we _always_ use `puppet agent --disable`,
so there is no reason we'd want to leave puppet agent stopped.