spec: merge repeated expects on same resource
We often run several consecutive expects on the same resource with
different .with() clauses. These may produce finer grained output only
if the wrapping `it` block has a title. Where the `it` blocks are
anonymous these have been merged. This is a small performance
enhancement, but it adds up over many consecutive calls.