Add DramaticIncreaseInHAProxyTraffic alert
This is replacing the HAProxyRunningCloseToBackendSessionLimit alert,
which did not measure what it said it measured, but did alert us to an
attack, so it wasn't totally useless either.

This alert should better reflect what is happening, and the decrease in
severity should make it clear that nothing is necessarily broken or
falling over if this is the only alert.