ci: run librarian-puppet w/o docker
This should run slightly faster than with docker as it allows us to use
a cached bundle. The main goal here is parity with the CI action,
which was also ported off of docker.

Rename workflow in librarian.yml for better identifcation in ui.

Use correct name of default branch so this actually runs on push.