Add exechosts -P for running a plan instead of a command
Bolt's inventory.yaml is perfectly fine for establishing a hierarchy of
hosts and configuring how to connect to them. But it really just does
not have an easy way to say "all the VMs not at ICTC" or "nonproduction
mysql physical servers."

And that's fine because someone already built taghosts, which works great.