stop rspec-puppet warnings about missing facts
rspec-puppet specifically checks that networking:ip and networking:fqdn
are present. If we override the networking fact in a test we must ensure
both of these child facts are present.

This can usually be accomplished by replacing merge with deep_merge,
but we must use identifiers (not strings) for the facts for this to work,
since that's how the default facts are structured.

Note: Removed many instances of `fqdn: "INVALID_DO_NOT_USE"` and
similar. This was a hack to stop the same warnings, but on an outdated
version of rspec-puppet.